+++ src/core/call.cs
@@ -12,6 +12,7 @@
  */
 
 using Fahrenheit.Atel;
+using Fahrenheit.FFX;
 
 namespace Fahrenheit;
 
@@ -199,13 +200,16 @@ public static unsafe partial class FhCall {
     // PUBLIC/UNRESTRICTED - BEGIN
 
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-    public delegate int CT_Exec(AtelBasicWorker* work, AtelStack* atelStack);
+    public delegate AeonAbilityRecipe CT_Init(AtelBasicWorker* work, int* storage, AtelStack* stack);
 
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-    public delegate int CT_RetInt(AtelBasicWorker* work, int* storage, AtelStack* atelStack);
+    public delegate int CT_Exec(AtelBasicWorker* work, AtelStack* stack);
 
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-    public delegate float CT_RetFloat(AtelBasicWorker* work, int* storage, AtelStack* atelStack);
+    public delegate int CT_RetInt(AtelBasicWorker* work, int* storage, AtelStack* stack);
+
+    [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
+    public delegate float CT_RetFloat(AtelBasicWorker* work, int* storage, AtelStack* stack);
 
     // PUBLIC/UNRESTRICTED - END
